Added some basic classes
Added Mixdac, DFB, and IFB.  DFB and IFB are Delay and Impulse,
respectively, with integrated feedback path. Just basic building blocks
for synthesizer graphs... I start to get insane if the synth gets too
large. Like gigantic functions.. I want to see the design at any level
as a single thing, much as possible.

Mixdac is the latest version of a global volume control and output stage
that I wrote for the CalArts ChucK class on Coursera. Useful for dealing
with multi-file projects using a the same audio output.