New class Automate
This is one I made for my week 7 project in Introduction to Programming for Musicians and Digital Performers (whew!). The idea here is a replacement for "automation" as you find in a DAW or a studio. You create a subclass of Automate, which overrides a pair of adj() methods, used to get/put value from something. Enable switch provided.